Updated the client to conform to the new API version, as of 02/02/2017

+++ Demo/src/main/java/nl/yurimeiburg/ondertekenen/demo/Demo.java
@@ -13,7 +13,6 @@ import org.springframework.context.support.ClassPathXmlApplicationContext;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.net.URISyntaxException;
 import java.util.concurrent.TimeUnit;
 
 
@@ -35,15 +34,17 @@ class Demo {
 
     /**
      * Entry point
+     *
      * @param args Startup parameters -- Not used.
      */
-    public static void main(String[] args){
+    public static void main(String[] args) {
         ApplicationContext context = new ClassPathXmlApplicationContext("spring.xml");
         new Demo(context.getBean("ondertekenenClient", OndertekenenClient.class));
     }
 
     /**
      * Create a new Demo Object, and start the flow of the Ondertekenen service.
+     *
      * @param ondertekenenClient - The DAO for the Ondertekenen service.
      */
     private Demo(OndertekenenClient ondertekenenClient) {
@@ -55,7 +56,7 @@ class Demo {
                         Helper.createDemoSigners()
                 ));
 
-        if(transaction == null)
+        if (transaction == null)
             throw new RuntimeException("Could not create a new transaction.");
 
         LOGGER.info("Created a new transaction, with ID: " + transaction.getId());
@@ -65,35 +66,44 @@ class Demo {
         LOGGER.info("Got the following transaction from server: " + (transaction.getId()));
 
         /* Upload a file! */
-        if(ondertekenenClient.uploadFile(transaction, Helper.getDemoFile())) {
+        if (ondertekenenClient.uploadFile(transaction, Helper.getDemoFile())) {
             LOGGER.info("Successfully uploaded PDF for transaction " + transaction.getId());
-        }else{
+        } else {
             LOGGER.error("Could not upload PDF. Deleting transaction: " + ondertekenenClient.deleteTransaction(transaction, true, "Testing."));
             return;
         }
 
+        /* Verify that the file is now in the transaction */
+        transaction = ondertekenenClient.getTransaction(transaction.getId());
+        LOGGER.info("Got the following transaction from server: " + (transaction));
+
+        /* Start the transaction */
+        ondertekenenClient.startTransaction(transaction.getId());
+
         /* Wait until the user has performed an action on the document */
-        blockUntilUserAction(ondertekenenClient, transaction);
+        transaction = blockUntilUserAction(ondertekenenClient, transaction);
 
         /* The user has either Signed or Rejected the transaction! Get the signed document + receipt. */
         LOGGER.info("Downloading proof.");
-        Document signedDocument = ondertekenenClient.getSignedDocument(transaction.getFile().getId(), true);
-        writeFile(signedDocument.getData(), SIGNED_DOCUMENT_OUTPUT_LOCATION);
-        Receipt receipt = ondertekenenClient.getReceipt(transaction.getId(), true);
-        writeFile(receipt.getData(), RECEIPT_OUTPUT_LOCATION);
-
+        if (TransactionStatus.REJECTED != transaction.getStatus()) { /* Can't download a non existing file */
+            Document signedDocument = ondertekenenClient.getSignedDocument(transaction.getId(), "Input.pdf", true);
+            writeFile(signedDocument.getData(), SIGNED_DOCUMENT_OUTPUT_LOCATION);
+            Receipt receipt = ondertekenenClient.getReceipt(transaction.getId(), true);
+            writeFile(receipt.getData(), RECEIPT_OUTPUT_LOCATION);
+        }
     }
 
     /**
      * Wait until the user has performed an action on the document
      * An action could be to sign it, or to reject it.
      * The polling interval is set in the variable POLLING_INTERVAL.
+     *
      * @param ondertekenenClient The client used to poll the new status
-     * @param transaction The transaction used to poll the status for.
+     * @param transaction        The transaction used to poll the status for.
      */
-    void blockUntilUserAction(OndertekenenClient ondertekenenClient, Transaction transaction){
+    Transaction blockUntilUserAction(OndertekenenClient ondertekenenClient, Transaction transaction) {
         /* Block until the user has signed: */
-        while(  transaction.getStatus() == TransactionStatus.WAITING_FOR_DOCUMENT ||
+        while (transaction.getStatus() == TransactionStatus.WAITING_FOR_DOCUMENT ||
                 transaction.getStatus() == TransactionStatus.WAITING_FOR_SIGNER ||
                 transaction.getStatus() == TransactionStatus.IN_PROGRESS) {
             try {
@@ -106,15 +116,17 @@ class Demo {
             transaction = ondertekenenClient.getTransaction(transaction.getId());
             LOGGER.info("Transaction is now in state: " + (transaction.getStatus()));
         }
+        return transaction;
     }
 
     /**
      * Small helper function to write a byte stream to a file.
+     *
      * @param data The byte stream to be written to file
      * @param file The output location
      */
-    void writeFile(byte[] data, File file){
-        try(FileOutputStream fos = new FileOutputStream(file)){
+    void writeFile(byte[] data, File file) {
+        try (FileOutputStream fos = new FileOutputStream(file)) {
             fos.write(data);
         } catch (IOException e) {
             LOGGER.error("Could not write to file: " + file.getName(), e);
